Split GroupValuesModal into these three files and add types for all the files.
